hey we can use that silly sign bit now

+++ standardBot/AllyData.java
@@ -7,7 +7,7 @@ public class AllyData {
 	public MapLocation location;
 	public int hp;
 	public boolean alive;
-	public AllyData(int memoryData) {
+	public AllyData(long memoryData) {
 		type = Utilities.bitInterval(memoryData, 28, 30);
 		location = Utilities.intToTarget(Utilities.bitInterval(memoryData, 10, 27));
 		hp = Utilities.bitInterval(memoryData, 1, 9);
@@ -20,21 +20,21 @@ public class AllyData {
 		hp = health;
 		alive = is_alive;
 	}
-	public int toInt(){
+	public long toInt(){
 		return type * (int) Math.pow(2, 28) + (int) Utilities.targetToInt(location) * (int) Math.pow(2, 10) + hp * 2 + ((alive) ? 1 : 0);
 	}
 	
-	public static boolean isAlive(int memoryData){
+	public static boolean isAlive(long memoryData){
 		//return Utilities.bitInterval(memoryData, 0, 0) == 1;
 		// 0 0 is the same as & 1
 		return (memoryData & 1) == 1;
 	}
 	
-	public static int getType(int memoryData) {
+	public static int getType(long memoryData) {
 		return Utilities.bitInterval(memoryData, 28, 30);
 	}
 	
-	public static int getLocInt(int memoryData){
+	public static int getLocInt(long memoryData){
 		return Utilities.bitInterval(memoryData, 10, 27);
 	}
 }
